combine GradeSellerModal.js with BuyItem.js
