+++ src/actions/VIPCheck.js
@@ -27,17 +27,19 @@ export const VIPCheckSpending = (userID) => {
 
 export const VIPCheckRating = (userID) => {
   const userURL = `users/${userID}`;
-
-  return database.ref(`${userURL}/rating`).once('value').then((snapshot) => {
+  database.ref(`${userURL}/rating`).once('value').then((snapshot) => {
     const RATING = snapshot.val();
-    console.log(`rating: ${RATING}`);
-
+    //console.log(`rating: ${RATING}`);
+ 
     if(RATING >= 4){
-      return database.ref(`${userURL}/rater_number`).once('value').then((snapshot) => {
-        const RATER_NUMBER = snapshot.val();
-        console.log(`raters: ${RATER_NUMBER}`);
-
-        if (RATER_NUMBER >= 3){
+      database.ref(`users/${userID}/grade`).once('value').then((snapshot) => {
+        let counter = 0;
+        const GRADES = snapshot.val();
+        for (let grade in GRADES){
+          counter += 1;
+        }
+        //console.log(`counter: ${counter}`);
+        if (counter >= 3){
           database.ref(`${userURL}/`).update({user_type: "VIP OU"});
         } else {
           database.ref(`${userURL}/`).update({user_type: "OU"});
